requirements.txt
pytz
aiohttp
aiodns
cchardet
ed25519
python-rapidjson
